added ggc support to maps and sequences
fixed optimizer deleting used tokens
added bool token type

+++ nest/error.h
@@ -76,8 +76,8 @@ extern "C" {
 
 typedef struct
 {
-    size_t line;
-    size_t col;
+    long line;
+    long col;
     char *filename;
     char *text;
     size_t text_len;
